src/requirements.txt
mysql-connector-python 
selenium